move obsolete algorithms from cudabgsegm to cudalegacy:
* GMG
* FGD
