+++ modules/cudabgsegm/perf/perf_bgsegm.cpp
@@ -42,10 +42,6 @@
 
 #include "perf_precomp.hpp"
 
-#ifdef HAVE_OPENCV_CUDAIMGPROC
-#  include "opencv2/cudaimgproc.hpp"
-#endif
-
 using namespace std;
 using namespace testing;
 using namespace perf;
@@ -61,83 +57,6 @@ using namespace perf;
 #  define BUILD_WITH_VIDEO_INPUT_SUPPORT 1
 #else
 #  define BUILD_WITH_VIDEO_INPUT_SUPPORT 0
-#endif
-
-//////////////////////////////////////////////////////
-// FGDStatModel
-
-#if BUILD_WITH_VIDEO_INPUT_SUPPORT
-
-DEF_PARAM_TEST_1(Video, string);
-
-PERF_TEST_P(Video, FGDStatModel,
-            Values(string("gpu/video/768x576.avi")))
-{
-    const int numIters = 10;
-
-    declare.time(60);
-
-    const string inputFile = perf::TestBase::getDataPath(GetParam());
-
-    cv::VideoCapture cap(inputFile);
-    ASSERT_TRUE(cap.isOpened());
-
-    cv::Mat frame;
-    cap >> frame;
-    ASSERT_FALSE(frame.empty());
-
-    if (PERF_RUN_CUDA())
-    {
-        cv::cuda::GpuMat d_frame(frame), foreground;
-
-        cv::Ptr<cv::cuda::BackgroundSubtractorFGD> d_fgd = cv::cuda::createBackgroundSubtractorFGD();
-        d_fgd->apply(d_frame, foreground);
-
-        int i = 0;
-
-        // collect performance data
-        for (; i < numIters; ++i)
-        {
-            cap >> frame;
-            ASSERT_FALSE(frame.empty());
-
-            d_frame.upload(frame);
-
-            startTimer();
-            if(!next())
-                break;
-
-            d_fgd->apply(d_frame, foreground);
-
-            stopTimer();
-        }
-
-        // process last frame in sequence to get data for sanity test
-        for (; i < numIters; ++i)
-        {
-            cap >> frame;
-            ASSERT_FALSE(frame.empty());
-
-            d_frame.upload(frame);
-
-            d_fgd->apply(d_frame, foreground);
-        }
-
-        CUDA_SANITY_CHECK(foreground, 1e-2, ERROR_RELATIVE);
-
-#ifdef HAVE_OPENCV_CUDAIMGPROC
-        cv::cuda::GpuMat background3, background;
-        d_fgd->getBackgroundImage(background3);
-        cv::cuda::cvtColor(background3, background, cv::COLOR_BGR2BGRA);
-        CUDA_SANITY_CHECK(background, 1e-2, ERROR_RELATIVE);
-#endif
-    }
-    else
-    {
-        FAIL_NO_CPU();
-    }
-}
-
 #endif
 
 //////////////////////////////////////////////////////
@@ -480,121 +399,6 @@ PERF_TEST_P(Video_Cn, MOG2GetBackgroundImage,
         TEST_CYCLE() mog2->getBackgroundImage(background);
 
         CPU_SANITY_CHECK(background);
-    }
-}
-
-#endif
-
-//////////////////////////////////////////////////////
-// GMG
-
-#if BUILD_WITH_VIDEO_INPUT_SUPPORT
-
-DEF_PARAM_TEST(Video_Cn_MaxFeatures, string, MatCn, int);
-
-PERF_TEST_P(Video_Cn_MaxFeatures, GMG,
-            Combine(Values(string("gpu/video/768x576.avi")),
-                    CUDA_CHANNELS_1_3_4,
-                    Values(20, 40, 60)))
-{
-    const int numIters = 150;
-
-    const std::string inputFile = perf::TestBase::getDataPath(GET_PARAM(0));
-    const int cn = GET_PARAM(1);
-    const int maxFeatures = GET_PARAM(2);
-
-    cv::VideoCapture cap(inputFile);
-    ASSERT_TRUE(cap.isOpened());
-
-    cv::Mat frame;
-    cap >> frame;
-    ASSERT_FALSE(frame.empty());
-
-    if (cn != 3)
-    {
-        cv::Mat temp;
-        if (cn == 1)
-            cv::cvtColor(frame, temp, cv::COLOR_BGR2GRAY);
-        else
-            cv::cvtColor(frame, temp, cv::COLOR_BGR2BGRA);
-        cv::swap(temp, frame);
-    }
-
-    if (PERF_RUN_CUDA())
-    {
-        cv::cuda::GpuMat d_frame(frame);
-        cv::cuda::GpuMat foreground;
-
-        cv::Ptr<cv::cuda::BackgroundSubtractorGMG> d_gmg = cv::cuda::createBackgroundSubtractorGMG();
-        d_gmg->setMaxFeatures(maxFeatures);
-
-        d_gmg->apply(d_frame, foreground);
-
-        int i = 0;
-
-        // collect performance data
-        for (; i < numIters; ++i)
-        {
-            cap >> frame;
-            if (frame.empty())
-            {
-                cap.release();
-                cap.open(inputFile);
-                cap >> frame;
-            }
-
-            if (cn != 3)
-            {
-                cv::Mat temp;
-                if (cn == 1)
-                    cv::cvtColor(frame, temp, cv::COLOR_BGR2GRAY);
-                else
-                    cv::cvtColor(frame, temp, cv::COLOR_BGR2BGRA);
-                cv::swap(temp, frame);
-            }
-
-            d_frame.upload(frame);
-
-            startTimer();
-            if(!next())
-                break;
-
-            d_gmg->apply(d_frame, foreground);
-
-            stopTimer();
-        }
-
-        // process last frame in sequence to get data for sanity test
-        for (; i < numIters; ++i)
-        {
-            cap >> frame;
-            if (frame.empty())
-            {
-                cap.release();
-                cap.open(inputFile);
-                cap >> frame;
-            }
-
-            if (cn != 3)
-            {
-                cv::Mat temp;
-                if (cn == 1)
-                    cv::cvtColor(frame, temp, cv::COLOR_BGR2GRAY);
-                else
-                    cv::cvtColor(frame, temp, cv::COLOR_BGR2BGRA);
-                cv::swap(temp, frame);
-            }
-
-            d_frame.upload(frame);
-
-            d_gmg->apply(d_frame, foreground);
-        }
-
-        CUDA_SANITY_CHECK(foreground);
-    }
-    else
-    {
-        FAIL_NO_CPU();
     }
 }
