+++ modules/cudalegacy/CMakeLists.txt
@@ -6,4 +6,4 @@ set(the_description "CUDA-accelerated Computer Vision (legacy)")
 
 ocv_warnings_disable(CMAKE_CXX_FLAGS /wd4127 /wd4130 /wd4324 /wd4512 /wd4310 -Wundef -Wmissing-declarations -Wuninitialized -Wshadow)
 
-ocv_define_module(cudalegacy opencv_core OPTIONAL opencv_objdetect)
+ocv_define_module(cudalegacy opencv_core opencv_video OPTIONAL opencv_objdetect opencv_imgproc opencv_cudaarithm opencv_cudafilters opencv_cudaimgproc)
