+++ modules/cudalegacy/src/cuda/fgd.cu
