+++ modules/cudalegacy/src/cuda/gmg.cu
