+++ modules/cudalegacy/src/gmg.cpp
